Move EnvironmentVariableDictionary to its own file.
