+++ src/EnvPrompt/Environment/Model/EnvironmentVariableDictionary.ts
@@ -1,0 +1,5 @@
+import { EnvironmentVariable } from '@/EnvPrompt/Environment/Model/EnvironmentVariable';
+
+export interface EnvironmentVariableDictionary {
+    [name: string]: EnvironmentVariable;
+}
